Moves `RemoteImageCommentsLoader` and collaborators to its own file
