+++ EssentialFeed/EssentialFeed/Comments/Feature/ImageComment.swift
@@ -1,0 +1,12 @@
+//
+//  Copyright © 2020 Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+public struct ImageComment: Equatable {
+    let id: UUID
+    let message: String
+    let createdAt: Date
+    let author: String
+}
